requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
django-tenants==3.7.0
Pillow==10.2.0
django-filter==23.5
django-guardian==2.4.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.27.1 